remove old ui
