+++ TravianOfficialCore/Parsers/VillageFields.cs
@@ -22,6 +22,7 @@ namespace TravianOfficialCore.Parsers
             if (string.IsNullOrEmpty(needClass)) return -1;
             var strResult = new string(needClass.Where(c => char.IsDigit(c)).ToArray());
             if (string.IsNullOrEmpty(strResult)) return -1;
+
             return int.Parse(strResult);
         }
