--- TravianOfficialNewHeroUICore/Parsers/StockBar.cs
@@ -1,120 +1,0 @@
-﻿using HtmlAgilityPack;
-using System.Linq;
-using System.Net;
-
-namespace TravianOfficialNewHeroUICore.Parsers
-{
-    public static class StockBar
-    {
-        public static long GetWood(HtmlDocument doc)
-        {
-            var woodNode = doc.GetElementbyId("l1");
-            if (woodNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(woodNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return long.Parse(valueStr);
-        }
-
-        public static long GetClay(HtmlDocument doc)
-        {
-            var clayNode = doc.GetElementbyId("l2");
-            if (clayNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(clayNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return long.Parse(valueStr);
-        }
-
-        public static int GetIron(HtmlDocument doc)
-        {
-            var ironNode = doc.GetElementbyId("l3");
-            if (ironNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(ironNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return int.Parse(valueStr);
-        }
-
-        public static long GetCrop(HtmlDocument doc)
-        {
-            var cropNode = doc.GetElementbyId("l4");
-            if (cropNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(cropNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return long.Parse(valueStr);
-        }
-
-        public static long GetFreeCrop(HtmlDocument doc)
-        {
-            var freecropNode = doc.GetElementbyId("stockBarFreeCrop");
-            if (freecropNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(freecropNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return long.Parse(valueStr);
-        }
-
-        public static long GetWarehouseCapacity(HtmlDocument doc)
-        {
-            var stockBarNode = doc.GetElementbyId("stockBar");
-            if (stockBarNode is null) return -1;
-            var warehouseNode = stockBarNode.Descendants("div").FirstOrDefault(x => x.HasClass("warehouse"));
-            if (warehouseNode is null) return -1;
-            var capacityNode = warehouseNode.Descendants("div").FirstOrDefault(x => x.HasClass("capacity"));
-            if (capacityNode is null) return -1;
-            var valueNode = capacityNode.Descendants("div").FirstOrDefault(x => x.HasClass("value"));
-            if (valueNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(valueNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return long.Parse(valueStr);
-        }
-
-        public static long GetGranaryCapacity(HtmlDocument doc)
-        {
-            var stockBarNode = doc.GetElementbyId("stockBar");
-            if (stockBarNode is null) return -1;
-            var granaryNode = stockBarNode.Descendants("div").FirstOrDefault(x => x.HasClass("granary"));
-            if (granaryNode is null) return -1;
-            var capacityNode = granaryNode.Descendants("div").FirstOrDefault(x => x.HasClass("capacity"));
-            if (capacityNode is null) return -1;
-            var valueNode = capacityNode.Descendants("div").FirstOrDefault(x => x.HasClass("value"));
-            if (valueNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(valueNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return long.Parse(valueStr);
-        }
-
-        public static int GetGold(HtmlDocument doc)
-        {
-            var goldNode = doc.DocumentNode.Descendants("div").FirstOrDefault(x => x.HasClass("ajaxReplaceableGoldAmount"));
-            if (goldNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(goldNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return int.Parse(valueStr);
-        }
-
-        public static int GetSilver(HtmlDocument doc)
-        {
-            var silverNode = doc.DocumentNode.Descendants("div").FirstOrDefault(x => x.HasClass("ajaxReplaceableSilverAmount"));
-            if (silverNode is null) return -1;
-            var valueStrFixed = WebUtility.HtmlDecode(silverNode.InnerText);
-            if (string.IsNullOrEmpty(valueStrFixed)) return -1;
-            var valueStr = new string(valueStrFixed.Where(c => char.IsDigit(c)).ToArray());
-            if (string.IsNullOrEmpty(valueStr)) return -1;
-            return int.Parse(valueStr);
-        }
-    }
-}
