--- TravianOfficialNewHeroUICore/TravianOfficialNewHeroUICore.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="HtmlAgilityPack" Version="1.11.43" />
-    <PackageReference Include="Selenium.Support" Version="4.3.0" />
-    <PackageReference Include="Selenium.WebDriver" Version="4.3.0" />
-  </ItemGroup>
-
-</Project>
